+++ Sources/PointFree/Session.swift
@@ -1,37 +1,8 @@
 import Foundation
 import HttpPipeline
+import Optics
 import Prelude
 import Tuple
-
-public struct Flash: Codable {
-  public enum Priority: String, Codable {
-    case error
-    case notice
-    case warning
-  }
-
-  public let priority: Priority
-  public let message: String
-}
-
-extension Flash: Equatable {
-  public static func ==(lhs: Flash, rhs: Flash) -> Bool {
-    return lhs.priority == rhs.priority && lhs.message == rhs.message
-  }
-}
-
-public struct Session: Codable {
-  public var flash: Flash?
-  public var userId: Database.User.Id?
-
-  public static let empty = Session(flash: nil, userId: nil)
-}
-
-extension Session: Equatable {
-  public static func ==(lhs: Session, rhs: Session) -> Bool {
-    return lhs.flash == rhs.flash && lhs.userId?.rawValue == rhs.userId?.rawValue
-  }
-}
 
 public func writeSessionCookieMiddleware<A>(_ update: @escaping (Session) -> Session)
   -> (Conn<HeadersOpen, A>)
@@ -53,6 +24,10 @@ public func writeSessionCookieMiddleware<A>(_ update: @escaping (Session) -> Ses
     }
 }
 
+public func flash<A>(_ priority: Flash.Priority, _ message: String) -> Middleware<HeadersOpen, HeadersOpen, A, A> {
+  return writeSessionCookieMiddleware(\.flash .~ Flash(priority: priority, message: message))
+}
+
 extension URLRequest {
   var session: Session {
     return self.cookies[pointFreeUserSession]
@@ -61,6 +36,36 @@ extension URLRequest {
           .verifiedValue(signedCookieValue: $0, secret: AppEnvironment.current.envVars.appSecret)
       }
       ?? .empty
+  }
+}
+
+public struct Session: Codable {
+  public var flash: Flash?
+  public var userId: Database.User.Id?
+
+  public static let empty = Session(flash: nil, userId: nil)
+}
+
+public struct Flash: Codable {
+  public enum Priority: String, Codable {
+    case error
+    case notice
+    case warning
+  }
+
+  public let priority: Priority
+  public let message: String
+}
+
+extension Session: Equatable {
+  public static func ==(lhs: Session, rhs: Session) -> Bool {
+    return lhs.flash == rhs.flash && lhs.userId?.rawValue == rhs.userId?.rawValue
+  }
+}
+
+extension Flash: Equatable {
+  public static func ==(lhs: Flash, rhs: Flash) -> Bool {
+    return lhs.priority == rhs.priority && lhs.message == rhs.message
   }
 }
